+++ Anarchy/Discord/Channel/ChannelExtensions.cs
@@ -6,9 +6,9 @@ namespace Discord
 {
     public static class ChannelExtensions
     {
-        public static Channel CreateChannel(this DiscordClient client, long guildId, ChannelProperties properties)
+        public static Channel CreateChannel(this DiscordClient client, long guildId, ChannelCreationProperties properties)
         {
-            var resp = client.HttpClient.PostAsync($"/guilds/{guildId}/channels", JsonConvert.SerializeObject(properties)).Result;
+            var resp = client.HttpClient.Post($"/guilds/{guildId}/channels", JsonConvert.SerializeObject(properties));
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new GuildNotFoundException(client, guildId);
@@ -18,9 +18,10 @@ namespace Discord
             return channel;
         }
 
-        public static Channel ModifyChannel(this DiscordClient client, long channelId, ChannelProperties properties)
+
+        public static Channel ModifyChannel(this DiscordClient client, long channelId, ChannelModProperties properties)
         {
-            var resp = client.HttpClient.PatchAsync($"/channels/{channelId}", JsonConvert.SerializeObject(properties)).Result;
+            var resp = client.HttpClient.Patch($"/channels/{channelId}", JsonConvert.SerializeObject(properties));
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new ChannelNotFoundException(client, channelId);
@@ -29,10 +30,11 @@ namespace Discord
             channel.Client = client;
             return channel;
         }
+
 
         public static Channel DeleteChannel(this DiscordClient client, long channelId)
         {
-            var resp = client.HttpClient.DeleteAsync($"/channels/{channelId}").Result;
+            var resp = client.HttpClient.Delete($"/channels/{channelId}");
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new ChannelNotFoundException(client, channelId);
@@ -42,9 +44,10 @@ namespace Discord
             return channel;
         }
 
+
         public static Channel GetChannel(this DiscordClient client, long channelId)
         {
-            var resp = client.HttpClient.GetAsync($"/channels/{channelId}").Result;
+            var resp = client.HttpClient.Get($"/channels/{channelId}");
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new ChannelNotFoundException(client, channelId);
@@ -52,11 +55,21 @@ namespace Discord
             Channel channel = JsonConvert.DeserializeObject<Channel>(resp.Content.ReadAsStringAsync().Result);
             channel.Client = client;
             return channel;
+        }
+
+        public static bool AddChannelPermissionOverwrite(this DiscordClient client, long channelId, PermissionOverwrite permission)
+        {
+            var resp = client.HttpClient.Put($"/channels/{channelId}/permissions/{permission.Id}");
+
+            if (resp.StatusCode == HttpStatusCode.NotFound)
+                throw new ChannelNotFoundException(client, channelId);
+
+            return resp.StatusCode == HttpStatusCode.NoContent;
         }
 
         public static List<Message> GetChannelMessages(this DiscordClient client, long channelId, int limit = 100, int afterId = 0)
         {
-            var resp = client.HttpClient.GetAsync($"/channels/{channelId}/messages?limit={limit}&after={afterId}").Result;
+            var resp = client.HttpClient.Get($"/channels/{channelId}/messages?limit={limit}&after={afterId}");
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new ChannelNotFoundException(client, channelId);
@@ -66,10 +79,11 @@ namespace Discord
             return messages;
         }
 
+
         #region pinning
         public static List<Message> GetChannelPinnedMessages(this DiscordClient client, long channelId)
         {
-            var resp = client.HttpClient.GetAsync($"/channels/{channelId}/pins").Result;
+            var resp = client.HttpClient.Get($"/channels/{channelId}/pins");
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new ChannelNotFoundException(client, channelId);
@@ -79,9 +93,10 @@ namespace Discord
             return messages;
         }
 
+
         public static bool PinChannelMessage(this DiscordClient client, long channelId, long messageId)
         {
-            var resp = client.HttpClient.PutAsync($"/channels/{channelId}/pins/{messageId}").Result;
+            var resp = client.HttpClient.Put($"/channels/{channelId}/pins/{messageId}");
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new MessageNotFoundException(client, messageId);
@@ -91,7 +106,7 @@ namespace Discord
 
         public static bool UnpinChannelMessage(this DiscordClient client, long channelId, long messageId)
         {
-            var resp = client.HttpClient.DeleteAsync($"/channels/{channelId}/pins/{messageId}").Result;
+            var resp = client.HttpClient.Delete($"/channels/{channelId}/pins/{messageId}");
 
             if (resp.StatusCode == HttpStatusCode.NotFound)
                 throw new MessageNotFoundException(client, messageId);
@@ -100,15 +115,17 @@ namespace Discord
         }
         #endregion
 
+
         public static Channel CreateDM(this DiscordClient client, long recipientId)
         {
-            var resp = client.HttpClient.PostAsync("/users/@me/channels", "{\"recipient_id\":\"" + recipientId + "\"}").Result;
+            var resp = client.HttpClient.Post("/users/@me/channels", "{\"recipient_id\":\"" + recipientId + "\"}");
             
             Channel channel = JsonConvert.DeserializeObject<Channel>(resp.Content.ReadAsStringAsync().Result);
             channel.Client = client;
             return channel;
         }
         
+
         public static Channel LeaveGroup(this DiscordClient client, long groupId)
         {
             return client.DeleteChannel(groupId);