+++ Anarchy/Discord/Guild/Models/BasicGuild.cs
@@ -12,12 +12,14 @@ namespace Discord
         public string Name { get; private set; }
 
         [JsonIgnore]
-        internal DiscordClient Client { get; set; }
+        internal virtual DiscordClient Client { get; set; }
+
 
         public bool Delete()
         {
             return Client.DeleteGuild(Id);
         }
+
 
         #region kick member
         public bool KickMember(long userId)
@@ -25,51 +27,61 @@ namespace Discord
             return Client.KickGuildMember(Id, userId);
         }
 
+
         public bool KickMember(User user)
         {
             return KickMember(user.Id);
         }
         #endregion
 
+
         public List<Channel> GetChannels()
         {
             return Client.GetGuildChannels(Id);
         }
+
 
         public virtual List<Reaction> GetReactions()
         {
             return Client.GetGuildReactions(Id);
         }
 
+
         public Reaction GetReaction(long reactionId)
         {
             return Client.GetGuildReaction(Id, reactionId);
         }
+
 
         public virtual List<Role> GetRoles()
         {
             return Client.GetGuildRoles(Id);
         }
 
+
         public List<Invite> GetInvites()
         {
             return Client.GetGuildInvites(Id);
         }
 
-        public Channel CreateChannel(ChannelProperties properties)
+
+        public Channel CreateChannel(ChannelCreationProperties properties)
         {
             return Client.CreateChannel(Id, properties);
         }
 
-        public Reaction CreateReaction(ReactionProperties properties)
+
+        public Reaction CreateReaction(ReactionCreationProperties properties)
         {
-            return Client.CreateReaction(Id, properties);
+            return Client.CreateGuildReaction(Id, properties);
         }
+
 
         public Role CreateRole()
         {
-            return Client.CreateGuildRole(Id);
+            return Client.CreateRole(Id);
         }
+
 
         public override string ToString()
         {