+++ Anarchy/Discord/Guild/Models/Guild.cs
@@ -20,11 +20,50 @@ namespace Discord
         [JsonProperty("owner_id")]
         public long OwnerId { get; private set; }
 
+        private List<Role> _roles;
         [JsonProperty("roles")]
-        public List<Role> Roles { get; private set; }
+        public List<Role> Roles
+        {
+            get { return _roles; }
+            private set
+            {
+                foreach (var role in value)
+                    role.GuildId = Id;
 
+                _roles = value;
+            }
+        }
+
+        private List<Reaction> _reactions;
         [JsonProperty("emojis")]
-        public List<Reaction> Reactions { get; private set; }
+        public List<Reaction> Reactions
+        {
+            get { return _reactions; }
+            private set
+            {
+                foreach (var reaction in value)
+                    reaction.GuildId = Id;
+
+                _reactions = value;
+            }
+        }
+
+        private DiscordClient _client;
+        internal override DiscordClient Client
+        {
+            get { return _client; }
+            set
+            {
+                _client = value;
+
+                foreach (var role in Roles)
+                    role.Client = Client;
+
+                foreach (var reaction in Reactions)
+                    reaction.Client = Client;
+            }
+        }
+
 
         public Guild Modify(GuildModProperties properties)
         {
@@ -39,11 +78,13 @@ namespace Discord
             return guild;
         }
 
+
         public override List<Role> GetRoles()
         {
             Roles = base.GetRoles();
             return Roles;
         }
+
 
         public override List<Reaction> GetReactions()
         {
@@ -51,14 +92,15 @@ namespace Discord
             return Reactions;
         }
 
+
         //TODO: Make this able to convert the bytes into an image (problems with the .webp extension i reckon)
         public byte[] DownloadIcon()
         {
             if (IconId == null)
                 return null;
                 
-            return Client.HttpClient.GetAsync($"https://cdn.discordapp.com/icons/{Id}/{IconId}.webp")
-                                    .Result.Content.ReadAsByteArrayAsync().Result;
+            return Client.HttpClient.Get($"https://cdn.discordapp.com/icons/{Id}/{IconId}.webp")
+                                    .Content.ReadAsByteArrayAsync().Result;
         }
     }
 }