+++ Anarchy/Discord/Reaction/Properties/ReactionModProperties.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Discord
+{
+    public class ReactionModProperties
+    {
+        [JsonProperty("name")]
+        public string Name { get; set; }
+    }
+}