+++ Anarchy/Discord/Relations/RelationsExtensions.cs
@@ -5,11 +5,13 @@ namespace Discord
 {
     public static class RelationsExtensions
     {
+        #region add friend
         public static bool AddFriend(this DiscordClient client, Recipient recipient)
         {
-            return client.HttpClient.PostAsync("/users/@me/relationships",
-                JsonConvert.SerializeObject(recipient)).Result.StatusCode == HttpStatusCode.NoContent;
+            return client.HttpClient.Post("/users/@me/relationships",
+                JsonConvert.SerializeObject(recipient)).StatusCode == HttpStatusCode.NoContent;
         }
+
 
         public static bool AddFriend(this DiscordClient client, string username, int discriminator)
         {
@@ -21,5 +23,24 @@ namespace Discord
 
             return client.AddFriend(recipient);
         }
+        #endregion
+
+
+        #region remove friend
+        public static bool RemoveFriend(this DiscordClient client, long userId)
+        {
+            var resp = client.HttpClient.Delete($"/users/@me/relationships/{userId}");
+
+            if (resp.StatusCode == HttpStatusCode.NotFound)
+                throw new UserNotFoundException(client, userId);
+
+            return resp.StatusCode == HttpStatusCode.NoContent;
+        }
+
+        public static bool RemoveFriend(this DiscordClient client, User user)
+        {
+            return client.RemoveFriend(user.Id);
+        }
+        #endregion
     }
 }